refactor: organize screens into logical subdirectories
- Group episode-related screens into src/screens/episode/
- Group medication-related screens into src/screens/medication/
- Group settings-related screens into src/screens/settings/
- Keep main screens (Dashboard, Analytics, etc.) at root level
- Update all imports in AppNavigator with clear grouping comments
- Fix all relative imports in moved screen files
- Update test imports to use new paths
- Maintain all existing functionality while improving organization

This follows the component-based pattern from EpisodeDetailScreen refactoring
and makes the codebase more maintainable and easier to navigate.
