+++ app/src/screens/__tests__/PerformanceScreen.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render, fireEvent, waitFor, act } from '@testing-library/react-native';
 import { Alert, Share } from 'react-native';
-import PerformanceScreen from '../PerformanceScreen';
+import PerformanceScreen from '../settings/PerformanceScreen';
 import {
   getPerformanceStats,
   clearPerformanceStats,
