+++ app/src/screens/episode/EditEpisodeNoteScreen.tsx
@@ -1,5 +1,5 @@
 import React, { useState, useEffect, useCallback } from 'react';
-import { logger } from '../utils/logger';
+import { logger } from '../../utils/logger';
 import {
   View,
   Text,
@@ -12,12 +12,12 @@ import {
   ActivityIndicator,
 } from 'react-native';
 import { NativeStackScreenProps } from '@react-navigation/native-stack';
-import { RootStackParamList } from '../navigation/types';
-import { episodeNoteRepository } from '../database/episodeRepository';
-import { EpisodeNote } from '../models/types';
+import { RootStackParamList } from '../../navigation/types';
+import { episodeNoteRepository } from '../../database/episodeRepository';
+import { EpisodeNote } from '../../models/types';
 import DateTimePicker from '@react-native-community/datetimepicker';
 import { format } from 'date-fns';
-import { useTheme, ThemeColors } from '../theme';
+import { useTheme, ThemeColors } from '../../theme';
 
 type Props = NativeStackScreenProps<RootStackParamList, 'EditEpisodeNote'>;
