+++ app/src/screens/episode/EditIntensityReadingScreen.tsx
@@ -1,5 +1,5 @@
 import React, { useState, useEffect, useCallback } from 'react';
-import { logger } from '../utils/logger';
+import { logger } from '../../utils/logger';
 import {
   View,
   Text,
@@ -11,13 +11,13 @@ import {
   ActivityIndicator,
 } from 'react-native';
 import { NativeStackScreenProps } from '@react-navigation/native-stack';
-import { RootStackParamList } from '../navigation/types';
-import { intensityRepository } from '../database/episodeRepository';
-import { IntensityReading } from '../models/types';
+import { RootStackParamList } from '../../navigation/types';
+import { intensityRepository } from '../../database/episodeRepository';
+import { IntensityReading } from '../../models/types';
 import DateTimePicker from '@react-native-community/datetimepicker';
 import { format } from 'date-fns';
-import { useTheme, ThemeColors } from '../theme';
-import { getPainColor, getPainLevel } from '../utils/painScale';
+import { useTheme, ThemeColors } from '../../theme';
+import { getPainColor, getPainLevel } from '../../utils/painScale';
 
 type Props = NativeStackScreenProps<RootStackParamList, 'EditIntensityReading'>;
