+++ app/src/screens/episode/EditSymptomLogScreen.tsx
@@ -7,8 +7,8 @@ import {
   Alert,
 } from 'react-native';
 import { NativeStackScreenProps } from '@react-navigation/native-stack';
-import { RootStackParamList } from '../navigation/types';
-import { useTheme, ThemeColors } from '../theme';
+import { RootStackParamList } from '../../navigation/types';
+import { useTheme, ThemeColors } from '../../theme';
 
 type Props = NativeStackScreenProps<RootStackParamList, 'EditSymptomLog'>;
